:recycle: refactor: mapper/converter 분리
