--- src/main/java/umc/product/domain/member/adviser/common/CommonMemberAdviser.java
@@ -1,42 +1,0 @@
-package umc.product.domain.member.adviser.common;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-import umc.product.domain.file.service.FileService;
-import umc.product.domain.member.dto.response.common.MemberIdResponse;
-import umc.product.domain.member.dto.response.common.MemberRoleResponse;
-import umc.product.domain.member.dto.response.common.MemberSearchResponse;
-import umc.product.domain.member.entity.Member;
-import umc.product.domain.member.mapper.MemberCodeMapper;
-import umc.product.domain.member.mapper.MemberMapper;
-import umc.product.domain.member.service.common.MemberService;
-
-@Component
-@RequiredArgsConstructor
-public class CommonMemberAdviser {
-    private final MemberService memberService;
-    private final FileService fileService;
-
-    private final MemberMapper memberMapper;
-    private final MemberCodeMapper memberCodeMapper;
-
-    public MemberRoleResponse verifyMemberCode(String code) {
-        return memberCodeMapper.toMemberRoleResponse(memberService.verifyMemberCode(code));
-    }
-    public MemberSearchResponse getMyProfile(Member member) {
-        return memberMapper.toSearchMemberResponse(member);
-    }
-    public MemberSearchResponse getProfile(Long memberId) {
-        Member member = memberService.findById(memberId);
-        return  memberMapper.toSearchMemberResponse(member);
-    }
-
-    public MemberIdResponse modifyMyProfileAvatar(Member member) {
-        //FileCreateResponse fileCreateResponse = fileService.createFile("AVATAR-IMAGE", file);
-        Member modifyMember = memberService.modifyMyProfileAvatar(member, "");
-        return  memberMapper.toMemberIdResponse(modifyMember.getId());
-    }
-
-
-
-}
