+++ src/main/java/umc/product/domain/member/adviser/member/MemberMemberAdviser.java
@@ -1,0 +1,40 @@
+package umc.product.domain.member.adviser.member;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+import umc.product.domain.file.service.FileService;
+import umc.product.domain.member.converter.response.MemberCodeConverter;
+import umc.product.domain.member.converter.response.MemberConverter;
+import umc.product.domain.member.dto.response.member.MemberIdResponse;
+import umc.product.domain.member.dto.response.member.MemberRoleResponse;
+import umc.product.domain.member.dto.response.member.MemberSearchResponse;
+import umc.product.domain.member.entity.Member;
+import umc.product.domain.member.service.common.MemberService;
+
+@Component
+@RequiredArgsConstructor
+public class MemberMemberAdviser {
+    private final MemberService memberService;
+    private final FileService fileService;
+
+    private final MemberConverter memberConverter;
+    private final MemberCodeConverter memberCodeConverter;
+
+    public MemberRoleResponse verifyMemberCode(String code) {
+        return memberCodeConverter.toMemberRoleResponse(memberService.verifyMemberCode(code));
+    }
+
+    public MemberSearchResponse getProfile(Long memberId) {
+        Member member = memberService.findById(memberId);
+        return  memberConverter.toSearchMemberResponse(member);
+    }
+
+    public MemberIdResponse modifyMyProfileAvatar(Member member) {
+        //FileCreateResponse fileCreateResponse = fileService.createFile("AVATAR-IMAGE", file);
+        Member modifyMember = memberService.modifyMyProfileAvatar(member, "");
+        return  memberConverter.toMemberIdResponse(modifyMember.getId());
+    }
+
+
+
+}
